rps-wagering/backend/requirements.txt
django
djangorestframework
channels
channels-redis
django-cors-headers
redis
web3
requests
